IMP - Refactor ATK skill hooks
